omicverse/external/saturn/requirements.txt
anndata
biopython
ensembl-rest
fair-esm
igraph
ipdb
ipython
louvain
matplotlib
notebook
numpy
pandas
plotly
scanpy
scikit-learn
scikit-learn-intelex
scipy
seaborn
record-keeper
tqdm
typed-argument-parser
faiss-gpu
torchvision
scvi-tools
scikit-misc